requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
openai==1.2.4
python-dotenv==1.0.0
sqlalchemy==2.0.22
bcrypt==4.0.1
email-validator==2.1.0
psycopg2-binary
